feat: record timestamp and play sound

+++ shared/src/commonMain/kotlin/data/features/trains/settings/SettingsRepository.kt
@@ -4,11 +4,11 @@ import com.russhwolf.settings.Settings
 import com.russhwolf.settings.set
 
 class SettingsRepository(private val settings: Settings) {
-    fun saveDefaultTrainTime(time: Int) {
+    fun saveDefaultTrainTime(time: Long) {
         settings[SETTINGS_TRAIN_TIME_KEY] = time
     }
 
-    fun getDefaultTrainTime(): Int = settings.getIntOrNull(SETTINGS_TRAIN_TIME_KEY) ?: 20
+    fun getDefaultTrainTime(): Long = settings.getLongOrNull(SETTINGS_TRAIN_TIME_KEY) ?: 20
 
     companion object {
         private const val SETTINGS_TRAIN_TIME_KEY = "settings_train_time_key"
