+++ shared/src/commonMain/kotlin/screens/train/settings/model/TrainSettingsEvent.kt
@@ -1,7 +1,7 @@
 package screens.train.settings.model
 
 sealed class TrainSettingsEvent {
-    data class TrainTimeChanged(val time: Int) : TrainSettingsEvent()
+    data class TrainTimeChanged(val time: Long) : TrainSettingsEvent()
     object ActionInvoked : TrainSettingsEvent()
     object OnBackClicked : TrainSettingsEvent()
     object OnSaveClicked : TrainSettingsEvent()
