+++ shared/src/commonMain/kotlin/screens/train/settings/model/TrainSettingsViewState.kt
@@ -1,5 +1,5 @@
 package screens.train.settings.model
 
 data class TrainSettingsViewState(
-    val trainTime: Int = 0,
+    val trainTime: Long = 0,
 )
